+++ src/scss/components/checkout/checkoutCart/_checkoutCart.scss
@@ -220,7 +220,7 @@
         display: flex;
         height: 28px;
         margin-bottom: 24px;
-        margin-top: -24px;
+        margin-top: 0;
 
         .banner-icon {
             align-items: center;
